Accept periods (full stops) in env values
The pattern used forbids usage of periods and quotations in env values.
While I can't judge on the usefulness of quotes within quotes at this
point (we'll need a real parser!), I do know that we have variables that
use full stops in there, and the pattern means we can't reload configs.

This patch makes it so instead of accepting all characters but ' and .,
only ' is excluded.